requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
backports-zoneinfo==0.2.1
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
boto3==1.34.87
    # via -r requirements/base.txt
botocore==1.34.118
    # via
    #   -r requirements/base.txt
    #   boto3
    #   s3transfer
certifi==2024.6.2
    # via
    #   -r requirements/base.txt
    #   requests
charset-normalizer==3.3.2
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.7
    # via code-annotations
code-annotations==1.8.0
    # via -r requirements/test.in
coverage[toml]==7.5.3
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-model-utils
    #   djangorestframework
django-model-utils==4.5.1
    # via -r requirements/base.txt
django-rest-framework==0.1.0
    # via -r requirements/base.txt
djangorestframework==3.15.1
    # via
    #   -r requirements/base.txt
    #   django-rest-framework
exceptiongroup==1.2.1
    # via pytest
idna==3.7
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via code-annotations
jmespath==1.0.1
    # via
    #   -r requirements/base.txt
    #   boto3
    #   botocore
markupsafe==2.1.5
    # via jinja2
openedx-atlas==0.6.0
    # via -r requirements/base.txt
packaging==24.0
    # via pytest
pbr==6.0.0
    # via stevedore
pluggy==1.5.0
    # via pytest
pytest==8.2.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/test.in
pytest-django==4.8.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   botocore
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.1
    # via code-annotations
requests==2.32.3
    # via -r requirements/base.txt
s3transfer==0.10.1
    # via
    #   -r requirements/base.txt
    #   boto3
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   python-dateutil
sqlparse==0.5.0
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.2.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.12.1
    # via
    #   -r requirements/base.txt
    #   asgiref
urllib3==1.26.18
    # via
    #   -r requirements/base.txt
    #   botocore
    #   requests
